+++ sdk/iothub/azure-resourcemanager-iothub/src/main/java/com/azure/resourcemanager/iothub/models/OperationInputs.java
@@ -6,14 +6,11 @@ package com.azure.resourcemanager.iothub.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /** Input values. */
 @Fluent
 public final class OperationInputs {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(OperationInputs.class);
-
     /*
      * The name of the IoT hub to check.
      */
@@ -47,9 +44,11 @@ public final class OperationInputs {
      */
     public void validate() {
         if (name() == null) {
-            throw logger
+            throw LOGGER
                 .logExceptionAsError(
                     new IllegalArgumentException("Missing required property name in model OperationInputs"));
         }
     }
+
+    private static final ClientLogger LOGGER = new ClientLogger(OperationInputs.class);
 }